requirements.txt
aiohttp
APScheduler
async-timeout
chardet
discord.py
emoji
multidict
python-dotenv
pytz
six
tzlocal
websockets
watchdog
